+++ frameworks/core/components_ng/pattern/scroll/inner/scroll_bar.h
@@ -472,22 +472,22 @@ public:
         dragFRCSceneCallback_ = std::move(dragFRCSceneCallback);
     }
 
-    double GetMainOffset(const Offset& offset) const
+    float GetMainOffset(const Offset& offset) const
     {
         return positionMode_ == PositionMode::BOTTOM ? offset.GetX() : offset.GetY();
     }
 
-    void SetDragStartPosition(double position)
+    void SetDragStartPosition(float position)
     {
         dragStartPosition_ = position;
     }
 
-    void SetDragEndPosition(double position)
+    void SetDragEndPosition(float position)
     {
         dragEndPosition_ = position;
     }
 
-    double GetDragOffset()
+    float GetDragOffset()
     {
         return dragEndPosition_ - dragStartPosition_;
     }
@@ -543,8 +543,8 @@ private:
     double barRegionSize_ = 0.0;
     double friction_ = BAR_FRICTION;
     double frictionPosition_ = 0.0;
-    double dragStartPosition_ = 0.0;
-    double dragEndPosition_ = 0.0;
+    float dragStartPosition_ = 0.0f;
+    float dragEndPosition_ = 0.0f;
 
     bool isScrollable_ = false;
