+++ frameworks/core/components_ng/pattern/scrollable/scrollable_properties.h
@@ -50,8 +50,11 @@ enum class ScrollSnapAlign {
 };
 
 enum class ScrollPagingStatus {
+    // no enablePaging and scrollSnap setting
     NONE = 0,
+    // scrollSnap has set, enablePaging is not effective
     INVALID,
+    // enablePaging is effective
     VALID,
 };
 
@@ -125,7 +128,7 @@ using OnScrollIndexEvent = std::function<void(int32_t, int32_t, int32_t)>;
 using ScrollPositionCallback = std::function<bool(double, int32_t source)>;
 using ScrollEndCallback = std::function<void()>;
 using CalePredictSnapOffsetCallback =
-                std::function<std::optional<float>(float delta, double dragDistance, double velocity)>;
+                std::function<std::optional<float>(float delta, float dragDistance, float velocity)>;
 using StartScrollSnapMotionCallback = std::function<void(float scrollSnapDelta, float scrollSnapVelocity)>;
 } // namespace OHOS::Ace
